#3 Changed where encrypted data was saved and combined mat_creation and mat_converter into one file and one class and renamed it Json_creation. Json files are created and saved in a directory called decrypted_dat under server
