+++ server/decryption/src/main.py
@@ -1,12 +1,9 @@
 # Import dependencies, libraries and classes
 import binary_extraction
 import folder_extraction
-import mat_creation
+import json_creation
 import copy_latest
 from threading import Timer
-import os
-import sys
-import mat_converter
 
 # Global Variables
 patient_name, date, measurements = "", "", {}
@@ -25,9 +22,10 @@ def main():
     # copy_latest.copy_files()
 
     # Extract all the data, index, settings and patient files into associated lists
-    folder = folder_extraction.FolderExtraction('../encrypted_data/')
+    folder = folder_extraction.FolderExtraction('data/')
     data_file_list, index_file_list, settings_file_list, qual_time_list, qual_str_list, patient_info = folder.extract_folders()
-    print(patient_info)
+    # print(patient_info)
+    print(data_file_list)
 
     # Extract the data, time_vector, units, start_time values and more from the binary files
     data = binary_extraction.DataExtraction(data_file_list,
@@ -38,6 +36,7 @@ def main():
                                             patient_info[0])
 
     patient_start, patient_discharged, patient_start_str, patient_name = data.extract_xml()
+    # print(patient_start_str)
     date = patient_start_str[0][:-4]
     index_dict = data.extract_index_files(patient_start)
     data_dict = data.extract_data_files()
@@ -45,15 +44,11 @@ def main():
     units, data_type, data_modality = data.extract_units()
     qual_str = data.extract_qual_string()
 
-    print(data_file_list)
+    # Create json files from the binary extracted values
+    json = json_creation.JsonCreation(qual_str, qual_dict, data_dict, patient_start_str, index_dict, units, data_file_list)
+    json.save_json()
 
-    # Create mat files from the binary extracted values
-    #mat = mat_creation.MatCreation(qual_str, qual_dict, data_dict, patient_start_str, index_dict, units, data_file_list)
     
-    #Convert mat files to json
-    #json_output = '../../decrypted_data'
-    #mat_convertor_object = mat_converter.MatConverter(mat)
-    #mat_converter_object.convert_to_json(mat, json_output)
 
 if __name__ == "__main__":
     main()
